[feat] add Exercise module
