--- src/post/post.repository.ts
@@ -1,7 +1,0 @@
-import { Repository, EntityRepository } from 'typeorm'
-import { Post } from './post.entity'
-
-@EntityRepository(Post)
-export class PostRepository extends Repository<Post> {
-
-}
