+++ src/post/post.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { getRepository, Repository } from "typeorm";
 
-import { Post, Exercise } from "./post.entity";
+import { Post } from "./post.entity";
 import { User } from "../user/user.entity";
 import { Like } from "../like/like.entity";
 
@@ -86,22 +86,6 @@ export class PostService {
   //     throw new Error(e);
   //   }
   // }
-
-  public async getExercise(context: object): Promise<{ Index: number[]; Name: string[] }> {
-    try {
-      const exercise: Exercise[] = await getRepository(Exercise)
-          .createQueryBuilder('e').select(['e.exerciseIndex', 'e.exerciseName']).getMany();
-      const exerciseIndex: number[] = [];
-      const name: string[] = [];
-      for (const node of exercise) {
-        exerciseIndex.push(node.exerciseIndex);
-        name.push(node.exerciseName);
-      }
-      return { Index: exerciseIndex, Name: name };
-    } catch(e) {
-      throw new Error(e);
-    }
-  }
 
   private async parseReturnData(data: Post[], userIndex: number): Promise<{ likeArray: number[]; PostData: PostInformation[] }> {
     const returnData: object[] = [];
